Add menu item to Settings to go to Cell Broadcast settings.
Add menu item to Settings when "config_cellBroadcastAppLinks"
config option is set to true to jump to the Cell Broadcast app
settings activity. This enables the Cell Broadcast launcher
icon to be removed, while still allowing access to the app
(by adding a similar menu item to the MMS app) and its settings
activity (this settings link).

The menu item will not be added if the CellBroadcastReceiver
package is disabled or not installed.

Bug: 6709985
Change-Id: If3a3016ceea70704790964c0b712c31cb38f7b62